Fix bug regarding sorting of speed records in ascending order according to time
